src/main/java/com/leetcode/p140/LinkedListCycle2142.java
package com.leetcode.p140;

import com.leetcode.ListNode;

/**
 * Created by sunilpatil on 5/11/17.
 */
public class LinkedListCycle2142 {
    public ListNode detectCycle(ListNode head) {
        ListNode slow = head;
        ListNode fast = head;
        while (fast.next != null && fast.next.next != null){
            slow = slow.next;
            fast = fast.next.next;
            if(slow == fast){
                slow = head;
                while (slow != fast){
                    slow = slow.next;
                    fast = fast.next;
                }
                return slow;
            }
        }
        return null;
    }
}